Move media source code from LiveTvManager to LiveTvMediaSourceProvider
